backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.4.0
pydantic==2.5.0
python-multipart==0.0.6
google-generativeai==0.3.2
scikit-learn==1.3.2
numpy==1.24.3
python-dotenv==1.0.0